Update minimal load of nuclear power to 0.8
Before, nuclear could not provide partload at all. With 0.8 we are giving nuclear some flexibility. We do not think that it matters much anyways, since in fact CAPEX of nuclear power is most important and it just gets higher in relative terms if nuclear power provides partload.

With 0.8 we also do not need to worry much about ramp rates as the max ramp rate by defintion will be 20%/hr which is maybe slightly too fast but not excessive.