Fix Slurm default resource disk_mb
When disk_mb is undefined, Snakemake adds the following line to the Slurm call:
> 'disk_mb=max(2*input.size_mb, 1000)'

It seems that older versions of Slurm have silently ignored this, but newer versions do not.
I don't know if that's true, but certainly the line was accepted on Euler in 2023 and now
it's not.

To fix this, I define a default resource of 1 GB.