Fix CSL issue
Today the joule.csl stopped working. When used, citeproc throws a "CiteprocParseError: No citation element present" error on all devices and in different networks. The file hasn't been changed so I assume this is a server error.

I replaced it with the CSL file for Cell Press generally, which, as far as I understand, is the same anyway.